remove eip2930 tx type (supposed won't be used)
